+++ apps/meteor/server/settings/accounts.ts
@@ -237,6 +237,7 @@ export const createAccountSettings = () =>
 		await this.add('Accounts_CustomFieldsToShowInUserInfo', '', {
 			type: 'string',
 			public: true,
+			i18nDescription: 'Accounts_CustomFieldsToShowInUserInfo_Description',
 		});
 		await this.add('Accounts_LoginExpiration', 90, {
 			type: 'int',
@@ -710,7 +711,7 @@ export const createAccountSettings = () =>
 				i18nLabel: 'Notification_volume',
 			});
 
-			await this.add('Accounts_Default_User_Preferences_voipRingerVolume', 100, {
+			await this.add('Accounts_Default_User_Preferences_callRingerVolume', 100, {
 				type: 'int',
 				public: true,
 				i18nLabel: 'Call_ringer_volume',