+++ packages/model-typings/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @rocket.chat/model-typings
 
+## 1.2.1
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/core-typings@7.2.1
+  </details>
+
 ## 1.2.0
 
 ### Minor Changes